Add a shared cache for the SEV certificate chain and mount it into the GHA runner
This directory is where all code paths will be updated to lead
to checking for cached certificate chains. This is motivated by
the need to have an SEV workflow that can scale hosting multiple
developers and potential CI jobs on the same machine. Right now,
there is an implicit "last person to cache wins" mode of operation
that is confusing and undocumented.

This patch lays the groundwork for our CI & developer machine to
*have* a shared directory to cache this certificate chain as well
as mount it into the GHA runner container so that the cached
certificate chain can be automatically updated by the CI when it
invalidates the cached certificate chain when it runs
dangerous_hw_tests.

The CI will always cache a valid certificate chain before and
after its runs. However, this patch does not solve the problem
of a developer running "dangerous_hw_tests" and forgetting to
cache a new chain. That will need to be addressed separately.

Signed-off-by: Connor Kuehl <ckuehl@redhat.com>